+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 - Misty step effect not applied if using Automated Animations as it has better built in support.
 - Quasit effects for Claws.
 - Invisibility feature effect (not spell), for monsters.
-- Ensure proficiency bonus to initiative is added to effects (Harengon)
+- Ensure proficiency bonus to initiative is added to effects (Harengon).
+- Improve checks for language modifiers that are not set, or shouldn't be present on DDB.
 
 # 3.4.57